Replace all uses of ContiguousArray with Array.
Swift's standard arrays do indeed use contiguous storage when storing
elements that are not @objc objects (in other words, the array is not
bridgeable to NSArray) and provide the same helper methods to access the
underlying storage unsafely. So, this change is a nice cleanup that uses
the more familiar array type.