Add a class that manages the logic for handling FastCGI requests.
This change also makes explicit the operation that writes a terminating
record to a FastCGI output stream, instead of rolling it into the flush
operation; the latter is unpredictable because flush can be called at
various times with no expectation of closing out the stream. This
termination must also be separate from the close operation on the
stream, because web servers can direct the application to not close
their streams.