Do not close wrapping I/O streams on deinit.
Some streams may not want to be closed on deinit, such as those
initialized with file descriptors. Having the wrapping streams
call close() would close them unconditionally and prematurely.
Instead, rely on the normal deinit calls to take care of this for
us; if a wrapped stream is deinitialized, eventually the underlying
stream will be too, and *it* will decide whether to close or not.

This also implies that you should not call close() on a wrapping
stream if you would not have called close() on the underlying stream,
either.